requirements.txt
Werkzeug==2.2.3
Flask==2.2.3
requests
google.generativeai
python-dotenv
openai
